requirements.txt
absl-py==0.9.0
aspy.yaml==1.3.0
backcall==0.1.0
cachetools==4.1.0
certifi==2020.4.5.1
cfgv==2.0.1
chardet==3.0.4
cycler==0.10.0
decorator==4.4.1
filterpy==1.4.5
future==0.18.2
google-auth==1.14.0
google-auth-oauthlib==0.4.1
grpcio==1.28.1
identify==1.4.11
idna==2.9
imageio==2.6.1
importlib-metadata==1.4.0
importlib-resources==1.0.2
ipdb==0.12.3
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
jedi==0.16.0
joblib==0.14.1
jupyter-client==6.0.0
jupyter-core==4.6.3
kiwisolver==1.1.0
llvmlite==0.31.0
Markdown==3.2.1
matplotlib==3.0.0
networkx==2.4
nodeenv==1.3.4
numba==0.47.0
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.1.2.30
parso==0.6.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.2.2
pre-commit==1.21.0
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.5.2
pyparsing==2.4.6
python-dateutil==2.8.1
PyWavelets==1.1.1
PyYAML==5.3
pyzmq==19.0.0
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-image==0.16.2
scikit-learn==0.22.1
scipy==1.1.0
six==1.14.0
tensorboard==2.2.1
tensorboard-plugin-wit==1.6.0.post3
toml==0.10.0
torch==1.2.0
torchvision==0.4.0
tornado==6.0.4
tqdm==4.26.0
traitlets==4.3.3
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.8
Werkzeug==1.0.1
yapf==0.29.0
zipp==2.1.0
